[Siddharth]Add:Refactor the code

+++ CommonLayer/Models/WishList/WishListModel.cs
@@ -6,13 +6,13 @@ namespace CommonLayer.Models.WishList
 {
     public class WishListModel
     {
-        public long WishListId { get; set; }
-        public long BookId { get; set; }
-        public long UserId { get; set; }
+        public int WishListId { get; set; }
+        public int BookId { get; set; }
+        public int UserId { get; set; }
         public string BookName { get; set; }
         public string AuthorName { get; set; }
-        public long OriginalPrice { get; set; }
-        public long DiscountPrice { get; set; }
+        public int OriginalPrice { get; set; }
+        public int DiscountPrice { get; set; }
         public string BookImage { get; set; }
     }
 }
